skill-b-pre-brief/requirements.txt
requests==2.31.0
python-dateutil==2.9.0
pytz==2024.1
pyyaml==6.0.1